Injecting key when keystore full fails.
New test with the above functionality was added.

+++ keystore/include/Key.hpp
@@ -15,5 +15,5 @@ public:
         return id != 0;
     }
 
-    uint8_t id = 0;
+    uint16_t id = 0;
 };
